Use NativeList in ApplicationVariables.cs instead. Add BuyerMeetsSeller function in "Functions - on triggers.cs".
Use NativeList in ApplicationVariables.cs instead.
Add BuyerMeetsSeller function in "Functions - on triggers.cs".

+++ Assets/LSW/Scripts/BodyPartIDToBodyPartInitialName.cs
@@ -4,7 +4,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_BodyPartIDToBodyPartInitialName {
 	
-		public int BodyPartID;
-		public int BodyPartInitialNameID;
+		public System.Int64  BodyPartID;
+		public System.Int64  BodyPartInitialNameID;
 	}
 }
