+++ Assets/LSW/Scripts/BodyPartInitialName.cs
@@ -2,8 +2,8 @@ using System.Text;
 
 namespace LSW{
 
-	[System.Serializable] public struct DataDefinition_BodyPartInitialName { public int ID; 
-		public int PlaceHolderTextID;
+	[System.Serializable] public struct DataDefinition_BodyPartInitialName { public System.Int64  ID; 
+		public System.Int64  PlaceHolderTextID;
 		
 	}
 }
