+++ Assets/LSW/Scripts/BuyerIDToBuyerInitialName.cs
@@ -2,7 +2,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_BuyerIDToBuyerInitialName
 	
-		{ public int ID; 
-		public int BuyerInitialNameID;
+		{ public System.Int64  ID; 
+		public System.Int64  BuyerInitialNameID;
 	}
 }
