+++ Assets/LSW/Scripts/BuyerInitialName.cs
@@ -4,8 +4,8 @@ namespace LSW
 {
 	[System.Serializable] public struct DataDefinition_BuyerInitialName
 	{
-		public int ID;
-		public int PlaceHolderTextID;
+		public System.Int64  ID;
+		public System.Int64  PlaceHolderTextID;
 		
 	}
 }
