+++ Assets/LSW/Scripts/BuyerPicture.cs
@@ -4,7 +4,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_BuyerPicture
 	
-		{ public int ID; 
+		{ public System.Int64  ID; 
 		public Sprite Instance_Picture;
 	}
 }
