+++ Assets/LSW/Scripts/DiscountIDToDiscountInDecimal.cs
@@ -2,7 +2,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_DiscountIDToDiscountInDecimal
 	
-		{ public int ID; 
-		public int DiscountInDecimalID;
+		{ public System.Int64  ID; 
+		public System.Int64  DiscountInDecimalID;
 	}
 }
