+++ Assets/LSW/Scripts/DiscountIDToDiscountInPercent.cs
@@ -2,7 +2,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_DiscountIDToDiscountInPercent
 	
-		{ public int ID; 
-		public int DiscountInPercentID;
+		{ public System.Int64  ID; 
+		public System.Int64  DiscountInPercentID;
 	}
 }
