+++ Assets/LSW/Scripts/EquipActionIDToBuyerID.cs
@@ -4,6 +4,6 @@ namespace LSW
 	[System.Serializable]
 	public struct DataDefinition_EquipActionIDToBuyerID
 	{
-		public int ID, BuyerID;
+		public System.Int64  ID, BuyerID;
 	}
 }
