+++ Assets/LSW/Scripts/EquipActionIDToProductID.cs
@@ -4,6 +4,6 @@ namespace LSW
 	[System.Serializable]
 	public struct DataDefinition_EquipActionIDToProductID
 	{
-		public int ID, ProductID;
+		public System.Int64  ID, ProductID;
 	}
 }
