+++ Assets/LSW/Scripts/HasIdentifier.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public class ParentDataDefinition_HasIdentifier {
 
-		public int ID; 
+		public System.Int64  ID; 
 	}
 }
