+++ Assets/LSW/Scripts/ItemIDToStock.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_ItemIDToStock
 	
-		{ public int ID,Stock;
+		{ public System.Int64  ID,Stock;
 	}
 }
