+++ Assets/LSW/Scripts/ItemSizeInitialName.cs
@@ -8,8 +8,8 @@ namespace LSW
 	public struct DataDefinition_ItemSizeInitialName
 
 	{
-		public int ID;
-		public int PlaceHolderTextID;
+		public System.Int64  ID;
+		public System.Int64  PlaceHolderTextID;
 
 	}
 }
