+++ Assets/LSW/Scripts/ItemTypeIDToItemXYZPositionID.cs
@@ -3,6 +3,6 @@ using System.Text;
 namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_ItemTypeIDToItemXYZPositionID  
-		{ public int ID, ItemXYZPositionID;
+		{ public System.Int64  ID, ItemXYZPositionID;
 	}
 }
