+++ Assets/LSW/Scripts/ItemTypeIDToLocationID.cs
@@ -3,6 +3,6 @@ using System.Text;
 namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_ItemTypeIDToLocationID 
-		{ public int ID, LocationID;
+		{ public System.Int64  ID, LocationID;
 	}
 }
