+++ Assets/LSW/Scripts/LocationID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_LocationID
 	
-		{ public int ID; 
+		{ public System.Int64  ID; 
 	}
 }
