+++ Assets/LSW/Scripts/LocationName.cs
@@ -4,7 +4,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_LocationName
 	
-		{ public int ID; 
-		public int PlaceHolderTextID;
+		{ public System.Int64  ID; 
+		public System.Int64  PlaceHolderTextID;
 	}
 }
