+++ Assets/LSW/Scripts/PictureIDToPictureDirectionTypeID.cs
@@ -2,7 +2,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_PictureIDToPictureDirectionTypeID
 	
-		{ public int ID; 
-		public int PictureDirectionTypeID;
+		{ public System.Int64  ID; 
+		public System.Int64  PictureDirectionTypeID;
 	}
 }
