+++ Assets/LSW/Scripts/PlaceHolderTextIDToTextID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_PlaceHolderTextIDToTextID
 	
-		{ public int ID, TextID;
+		{ public System.Int64  ID, TextID;
 	}
 }
