+++ Assets/LSW/Scripts/PriceIDToPriceInDecimal.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_PriceIDToPriceInDecimal
 	
-		{ public int ID, PriceInDecimalID;
+		{ public System.Int64  ID, PriceInDecimalID;
 	}
 }
