+++ Assets/LSW/Scripts/SellerIDToStoreID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_SellerIDToStoreID
 	 
-		{ public int ID, StoreID;
+		{ public System.Int64  ID, StoreID;
 	}
 }
