+++ Assets/LSW/Scripts/StoreIDToLocationID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_StoreIDToLocationID
 	
-		{ public int ID, LocationID;
+		{ public System.Int64  ID, LocationID;
 	}
 }
