+++ Assets/LSW/Scripts/StoreInitialName.cs
@@ -7,8 +7,8 @@ namespace LSW
 	public struct DataDefinition_StoreInitialName
 
 	{
-		public int ID;
-		public int PlaceHolderTextID;
+		public System.Int64  ID;
+		public System.Int64  PlaceHolderTextID;
 
 	}
 }
