+++ Assets/LSW/Scripts/TextIDToLanguageID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TextIDToLanguageID
 	
-		{ public int ID, LanguageID;
+		{ public System.Int64  ID, LanguageID;
 	}
 }
