+++ Assets/LSW/Scripts/TransactionIDToBuyerID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TransactionIDToBuyerID
 	
-		{ public int ID, BuyerID;
+		{ public System.Int64  ID, BuyerID;
 	}
 }
