+++ Assets/LSW/Scripts/TransactionIDToDate.cs
@@ -4,7 +4,7 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TransactionIDToDate
 	
-		{ public int ID; 
+		{ public System.Int64  ID; 
 		public DateTime DateTimeInstance;
 	}
 }
