+++ Assets/LSW/Scripts/TransactionIDToProductID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TransactionIDToProductID
 	
-		{ public int ID, ProductID;
+		{ public System.Int64  ID, ProductID;
 	}
 }
