+++ Assets/LSW/Scripts/TransactionIDToSellerID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TransactionIDToSellerID
 	
-		{ public int ID, SellerID;
+		{ public System.Int64  ID, SellerID;
 	}
 }
