+++ Assets/LSW/Scripts/TransactionIDToShopkeeperID.cs
@@ -2,6 +2,6 @@ namespace LSW{
 
 	[System.Serializable] public struct DataDefinition_TransactionIDToShopkeeperID
 	
-		{ public int ID, ShopkeeperID;
+		{ public System.Int64  ID, ShopkeeperID;
 	}
 }
